refactor: move fetchData and shuffleProfiles to helpers folder
