+++ src/helpers/fetchData.js
@@ -1,0 +1,10 @@
+export const fetchData = async (file) => {
+  try {
+    const response = await fetch(file);
+    const data = await response.json();
+    return data;
+  } catch (error) {
+    console.error('Error fetching data:', error);
+    return [];
+  }
+};
